+++ packages/manager/apps/web/client/app/app.js
@@ -34,6 +34,7 @@ import ovhNotificationsSidebar from '@ovh-ux/manager-notifications-sidebar';
 import ovhManagerAccountMigration from '@ovh-ux/manager-account-migration';
 import ovhManagerBanner from '@ovh-ux/manager-banner';
 import ovhManagerCatalogPrice from '@ovh-ux/manager-catalog-price';
+import ovhManagerIncidentBanner from '@ovh-ux/manager-incident-banner';
 import ovhManagerNavbar from '@ovh-ux/manager-navbar';
 import ovhManagerProductOffers from '@ovh-ux/manager-product-offers';
 import uiRouter, { RejectType } from '@uirouter/angularjs';
@@ -125,6 +126,7 @@ angular
       ovhManagerAccountMigration,
       ovhManagerBanner,
       ovhManagerCatalogPrice,
+      ovhManagerIncidentBanner,
       ovhManagerNavbar,
       ovhManagerProductOffers,
       ovhNotificationsSidebar,