+++ packages/manager/apps/web/CHANGELOG.md
@@ -1,3 +1,12 @@
+# [15.17.0](https://github.com/ovh/manager/compare/@ovh-ux/manager-web@15.16.8...@ovh-ux/manager-web@15.17.0) (2021-03-17)
+
+
+### Features
+
+* add incident migration banner ([f05c541](https://github.com/ovh/manager/commit/f05c541b027a1af92f211c7d17bb08c8c466eb50))
+
+
+
 ## [15.16.8](https://github.com/ovh/manager/compare/@ovh-ux/manager-web@15.16.7...@ovh-ux/manager-web@15.16.8) (2021-03-16)