+++ packages/manager/modules/incident-banner/CHANGELOG.md
@@ -1,3 +1,12 @@
+# [1.1.0](https://github.com/ovh/manager/compare/@ovh-ux/manager-incident-banner@1.0.0...@ovh-ux/manager-incident-banner@1.1.0) (2021-03-17)
+
+
+### Features
+
+* add incident migration banner ([f05c541](https://github.com/ovh/manager/commit/f05c541b027a1af92f211c7d17bb08c8c466eb50))
+
+
+
 # [1.0.0](https://github.com/ovh/manager/compare/@ovh-ux/manager-incident-banner@0.0.0...@ovh-ux/manager-incident-banner@1.0.0) (2021-03-14)